Extract lookup table initialization to lookup.go
